LegacyCode/requirements.txt
setuptools
tequila-basic==1.0.0
Deprecated
